+++ enterprise/internal/codeintel/autoindexing/internal/store/store.go
@@ -21,6 +21,8 @@ type Store interface {
 	SetInferenceScript(ctx context.Context, script string) error
 
 	// Repository configuration
+	RepositoryExceptions(ctx context.Context, repositoryID int) (canSchedule, canInfer bool, _ error)
+	SetRepositoryExceptions(ctx context.Context, repositoryID int, canSchedule, canInfer bool) error
 	GetIndexConfigurationByRepositoryID(ctx context.Context, repositoryID int) (shared.IndexConfiguration, bool, error)
 	UpdateIndexConfigurationByRepositoryID(ctx context.Context, repositoryID int, data []byte) error