+++ enterprise/internal/codeintel/autoindexing/mocks_test.go
@@ -61,6 +61,9 @@ type MockStore struct {
 	// QueueRepoRevFunc is an instance of a mock function object controlling
 	// the behavior of the method QueueRepoRev.
 	QueueRepoRevFunc *StoreQueueRepoRevFunc
+	// RepositoryExceptionsFunc is an instance of a mock function object
+	// controlling the behavior of the method RepositoryExceptions.
+	RepositoryExceptionsFunc *StoreRepositoryExceptionsFunc
 	// RepositoryIDsWithConfigurationFunc is an instance of a mock function
 	// object controlling the behavior of the method
 	// RepositoryIDsWithConfiguration.
@@ -71,6 +74,9 @@ type MockStore struct {
 	// SetInferenceScriptFunc is an instance of a mock function object
 	// controlling the behavior of the method SetInferenceScript.
 	SetInferenceScriptFunc *StoreSetInferenceScriptFunc
+	// SetRepositoryExceptionsFunc is an instance of a mock function object
+	// controlling the behavior of the method SetRepositoryExceptions.
+	SetRepositoryExceptionsFunc *StoreSetRepositoryExceptionsFunc
 	// TopRepositoriesToConfigureFunc is an instance of a mock function
 	// object controlling the behavior of the method
 	// TopRepositoriesToConfigure.
@@ -147,6 +153,11 @@ func NewMockStore() *MockStore {
 				return
 			},
 		},
+		RepositoryExceptionsFunc: &StoreRepositoryExceptionsFunc{
+			defaultHook: func(context.Context, int) (r0 bool, r1 bool, r2 error) {
+				return
+			},
+		},
 		RepositoryIDsWithConfigurationFunc: &StoreRepositoryIDsWithConfigurationFunc{
 			defaultHook: func(context.Context, int, int) (r0 []shared1.RepositoryWithAvailableIndexers, r1 int, r2 error) {
 				return
@@ -159,6 +170,11 @@ func NewMockStore() *MockStore {
 		},
 		SetInferenceScriptFunc: &StoreSetInferenceScriptFunc{
 			defaultHook: func(context.Context, string) (r0 error) {
+				return
+			},
+		},
+		SetRepositoryExceptionsFunc: &StoreSetRepositoryExceptionsFunc{
+			defaultHook: func(context.Context, int, bool, bool) (r0 error) {
 				return
 			},
 		},
@@ -244,6 +260,11 @@ func NewStrictMockStore() *MockStore {
 				panic("unexpected invocation of MockStore.QueueRepoRev")
 			},
 		},
+		RepositoryExceptionsFunc: &StoreRepositoryExceptionsFunc{
+			defaultHook: func(context.Context, int) (bool, bool, error) {
+				panic("unexpected invocation of MockStore.RepositoryExceptions")
+			},
+		},
 		RepositoryIDsWithConfigurationFunc: &StoreRepositoryIDsWithConfigurationFunc{
 			defaultHook: func(context.Context, int, int) ([]shared1.RepositoryWithAvailableIndexers, int, error) {
 				panic("unexpected invocation of MockStore.RepositoryIDsWithConfiguration")
@@ -257,6 +278,11 @@ func NewStrictMockStore() *MockStore {
 		SetInferenceScriptFunc: &StoreSetInferenceScriptFunc{
 			defaultHook: func(context.Context, string) error {
 				panic("unexpected invocation of MockStore.SetInferenceScript")
+			},
+		},
+		SetRepositoryExceptionsFunc: &StoreSetRepositoryExceptionsFunc{
+			defaultHook: func(context.Context, int, bool, bool) error {
+				panic("unexpected invocation of MockStore.SetRepositoryExceptions")
 			},
 		},
 		TopRepositoriesToConfigureFunc: &StoreTopRepositoriesToConfigureFunc{
@@ -319,6 +345,9 @@ func NewMockStoreFrom(i store.Store) *MockStore {
 		QueueRepoRevFunc: &StoreQueueRepoRevFunc{
 			defaultHook: i.QueueRepoRev,
 		},
+		RepositoryExceptionsFunc: &StoreRepositoryExceptionsFunc{
+			defaultHook: i.RepositoryExceptions,
+		},
 		RepositoryIDsWithConfigurationFunc: &StoreRepositoryIDsWithConfigurationFunc{
 			defaultHook: i.RepositoryIDsWithConfiguration,
 		},
@@ -327,6 +356,9 @@ func NewMockStoreFrom(i store.Store) *MockStore {
 		},
 		SetInferenceScriptFunc: &StoreSetInferenceScriptFunc{
 			defaultHook: i.SetInferenceScript,
+		},
+		SetRepositoryExceptionsFunc: &StoreSetRepositoryExceptionsFunc{
+			defaultHook: i.SetRepositoryExceptions,
 		},
 		TopRepositoriesToConfigureFunc: &StoreTopRepositoriesToConfigureFunc{
 			defaultHook: i.TopRepositoriesToConfigure,
@@ -1570,6 +1602,117 @@ func (c StoreQueueRepoRevFuncCall) Results() []interface{} {
 	return []interface{}{c.Result0}
 }
 
+// StoreRepositoryExceptionsFunc describes the behavior when the
+// RepositoryExceptions method of the parent MockStore instance is invoked.
+type StoreRepositoryExceptionsFunc struct {
+	defaultHook func(context.Context, int) (bool, bool, error)
+	hooks       []func(context.Context, int) (bool, bool, error)
+	history     []StoreRepositoryExceptionsFuncCall
+	mutex       sync.Mutex
+}
+
+// RepositoryExceptions delegates to the next hook function in the queue and
+// stores the parameter and result values of this invocation.
+func (m *MockStore) RepositoryExceptions(v0 context.Context, v1 int) (bool, bool, error) {
+	r0, r1, r2 := m.RepositoryExceptionsFunc.nextHook()(v0, v1)
+	m.RepositoryExceptionsFunc.appendCall(StoreRepositoryExceptionsFuncCall{v0, v1, r0, r1, r2})
+	return r0, r1, r2
+}
+
+// SetDefaultHook sets function that is called when the RepositoryExceptions
+// method of the parent MockStore instance is invoked and the hook queue is
+// empty.
+func (f *StoreRepositoryExceptionsFunc) SetDefaultHook(hook func(context.Context, int) (bool, bool, error)) {
+	f.defaultHook = hook
+}
+
+// PushHook adds a function to the end of hook queue. Each invocation of the
+// RepositoryExceptions method of the parent MockStore instance invokes the
+// hook at the front of the queue and discards it. After the queue is empty,
+// the default hook function is invoked for any future action.
+func (f *StoreRepositoryExceptionsFunc) PushHook(hook func(context.Context, int) (bool, bool, error)) {
+	f.mutex.Lock()
+	f.hooks = append(f.hooks, hook)
+	f.mutex.Unlock()
+}
+
+// SetDefaultReturn calls SetDefaultHook with a function that returns the
+// given values.
+func (f *StoreRepositoryExceptionsFunc) SetDefaultReturn(r0 bool, r1 bool, r2 error) {
+	f.SetDefaultHook(func(context.Context, int) (bool, bool, error) {
+		return r0, r1, r2
+	})
+}
+
+// PushReturn calls PushHook with a function that returns the given values.
+func (f *StoreRepositoryExceptionsFunc) PushReturn(r0 bool, r1 bool, r2 error) {
+	f.PushHook(func(context.Context, int) (bool, bool, error) {
+		return r0, r1, r2
+	})
+}
+
+func (f *StoreRepositoryExceptionsFunc) nextHook() func(context.Context, int) (bool, bool, error) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
+
+	if len(f.hooks) == 0 {
+		return f.defaultHook
+	}
+
+	hook := f.hooks[0]
+	f.hooks = f.hooks[1:]
+	return hook
+}
+
+func (f *StoreRepositoryExceptionsFunc) appendCall(r0 StoreRepositoryExceptionsFuncCall) {
+	f.mutex.Lock()
+	f.history = append(f.history, r0)
+	f.mutex.Unlock()
+}
+
+// History returns a sequence of StoreRepositoryExceptionsFuncCall objects
+// describing the invocations of this function.
+func (f *StoreRepositoryExceptionsFunc) History() []StoreRepositoryExceptionsFuncCall {
+	f.mutex.Lock()
+	history := make([]StoreRepositoryExceptionsFuncCall, len(f.history))
+	copy(history, f.history)
+	f.mutex.Unlock()
+
+	return history
+}
+
+// StoreRepositoryExceptionsFuncCall is an object that describes an
+// invocation of method RepositoryExceptions on an instance of MockStore.
+type StoreRepositoryExceptionsFuncCall struct {
+	// Arg0 is the value of the 1st argument passed to this method
+	// invocation.
+	Arg0 context.Context
+	// Arg1 is the value of the 2nd argument passed to this method
+	// invocation.
+	Arg1 int
+	// Result0 is the value of the 1st result returned from this method
+	// invocation.
+	Result0 bool
+	// Result1 is the value of the 2nd result returned from this method
+	// invocation.
+	Result1 bool
+	// Result2 is the value of the 3rd result returned from this method
+	// invocation.
+	Result2 error
+}
+
+// Args returns an interface slice containing the arguments of this
+// invocation.
+func (c StoreRepositoryExceptionsFuncCall) Args() []interface{} {
+	return []interface{}{c.Arg0, c.Arg1}
+}
+
+// Results returns an interface slice containing the results of this
+// invocation.
+func (c StoreRepositoryExceptionsFuncCall) Results() []interface{} {
+	return []interface{}{c.Result0, c.Result1, c.Result2}
+}
+
 // StoreRepositoryIDsWithConfigurationFunc describes the behavior when the
 // RepositoryIDsWithConfiguration method of the parent MockStore instance is
 // invoked.
@@ -1901,6 +2044,118 @@ func (c StoreSetInferenceScriptFuncCall) Args() []interface{} {
 // Results returns an interface slice containing the results of this
 // invocation.
 func (c StoreSetInferenceScriptFuncCall) Results() []interface{} {
+	return []interface{}{c.Result0}
+}
+
+// StoreSetRepositoryExceptionsFunc describes the behavior when the
+// SetRepositoryExceptions method of the parent MockStore instance is
+// invoked.
+type StoreSetRepositoryExceptionsFunc struct {
+	defaultHook func(context.Context, int, bool, bool) error
+	hooks       []func(context.Context, int, bool, bool) error
+	history     []StoreSetRepositoryExceptionsFuncCall
+	mutex       sync.Mutex
+}
+
+// SetRepositoryExceptions delegates to the next hook function in the queue
+// and stores the parameter and result values of this invocation.
+func (m *MockStore) SetRepositoryExceptions(v0 context.Context, v1 int, v2 bool, v3 bool) error {
+	r0 := m.SetRepositoryExceptionsFunc.nextHook()(v0, v1, v2, v3)
+	m.SetRepositoryExceptionsFunc.appendCall(StoreSetRepositoryExceptionsFuncCall{v0, v1, v2, v3, r0})
+	return r0
+}
+
+// SetDefaultHook sets function that is called when the
+// SetRepositoryExceptions method of the parent MockStore instance is
+// invoked and the hook queue is empty.
+func (f *StoreSetRepositoryExceptionsFunc) SetDefaultHook(hook func(context.Context, int, bool, bool) error) {
+	f.defaultHook = hook
+}
+
+// PushHook adds a function to the end of hook queue. Each invocation of the
+// SetRepositoryExceptions method of the parent MockStore instance invokes
+// the hook at the front of the queue and discards it. After the queue is
+// empty, the default hook function is invoked for any future action.
+func (f *StoreSetRepositoryExceptionsFunc) PushHook(hook func(context.Context, int, bool, bool) error) {
+	f.mutex.Lock()
+	f.hooks = append(f.hooks, hook)
+	f.mutex.Unlock()
+}
+
+// SetDefaultReturn calls SetDefaultHook with a function that returns the
+// given values.
+func (f *StoreSetRepositoryExceptionsFunc) SetDefaultReturn(r0 error) {
+	f.SetDefaultHook(func(context.Context, int, bool, bool) error {
+		return r0
+	})
+}
+
+// PushReturn calls PushHook with a function that returns the given values.
+func (f *StoreSetRepositoryExceptionsFunc) PushReturn(r0 error) {
+	f.PushHook(func(context.Context, int, bool, bool) error {
+		return r0
+	})
+}
+
+func (f *StoreSetRepositoryExceptionsFunc) nextHook() func(context.Context, int, bool, bool) error {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
+
+	if len(f.hooks) == 0 {
+		return f.defaultHook
+	}
+
+	hook := f.hooks[0]
+	f.hooks = f.hooks[1:]
+	return hook
+}
+
+func (f *StoreSetRepositoryExceptionsFunc) appendCall(r0 StoreSetRepositoryExceptionsFuncCall) {
+	f.mutex.Lock()
+	f.history = append(f.history, r0)
+	f.mutex.Unlock()
+}
+
+// History returns a sequence of StoreSetRepositoryExceptionsFuncCall
+// objects describing the invocations of this function.
+func (f *StoreSetRepositoryExceptionsFunc) History() []StoreSetRepositoryExceptionsFuncCall {
+	f.mutex.Lock()
+	history := make([]StoreSetRepositoryExceptionsFuncCall, len(f.history))
+	copy(history, f.history)
+	f.mutex.Unlock()
+
+	return history
+}
+
+// StoreSetRepositoryExceptionsFuncCall is an object that describes an
+// invocation of method SetRepositoryExceptions on an instance of MockStore.
+type StoreSetRepositoryExceptionsFuncCall struct {
+	// Arg0 is the value of the 1st argument passed to this method
+	// invocation.
+	Arg0 context.Context
+	// Arg1 is the value of the 2nd argument passed to this method
+	// invocation.
+	Arg1 int
+	// Arg2 is the value of the 3rd argument passed to this method
+	// invocation.
+	Arg2 bool
+	// Arg3 is the value of the 4th argument passed to this method
+	// invocation.
+	Arg3 bool
+	// Result0 is the value of the 1st result returned from this method
+	// invocation.
+	Result0 error
+}
+
+// Args returns an interface slice containing the arguments of this
+// invocation.
+func (c StoreSetRepositoryExceptionsFuncCall) Args() []interface{} {
+	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3}
+}
+
+// Results returns an interface slice containing the results of this
+// invocation.
+func (c StoreSetRepositoryExceptionsFuncCall) Results() []interface{} {
 	return []interface{}{c.Result0}
 }