+++ migrations/frontend/1683561153_add_autoindexing_repo_exceptions_table/down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS codeintel_autoindexing_exceptions;